+++ src/Component/Action/index.jsx
@@ -1,7 +1,9 @@
 import React, { useState } from "react";
 
 import { IoLogoYoutube } from "react-icons/io";
-import MyModal from "../MyModal";
+import Section from "../ui/section";
+import FeatureHeading from "../ui/featureheading";
+import MyModal from "../modals/video";
 
 import classes from "./index.module.css";
 
@@ -11,17 +13,25 @@ const Action = () => {
   const closeModal = () => setShowModal(false);
 
   return (
-    <div className={classes["action"]}>
-      <img
-        className={classes["action__image"]}
-        src="/Images/action img.jpeg"
-        alt="cloth with button icon "
+    <Section background="green">
+      <FeatureHeading
+        feature="View in Action"
+        heading="white"
+        font="biggerfont"
+        border="bordernon"
       />
-      <div className={classes["action__icon"]}>
-        <IoLogoYoutube onClick={() => setShowModal(true)}></IoLogoYoutube>
-        {showModal && <MyModal closeModal={closeModal} />}
+      <div className={classes["action"]}>
+        <img
+          className={classes["action__image"]}
+          src="/Images/action img.jpeg"
+          alt="cloth with button icon "
+        />
+        <div className={classes["action__icon"]}>
+          <IoLogoYoutube onClick={() => setShowModal(true)}></IoLogoYoutube>
+          {showModal && <MyModal closeModal={closeModal} />}
+        </div>
       </div>
-    </div>
+    </Section>
   );
 };
