--- src/Component/HomePage/index.css
@@ -1,11 +1,0 @@
-.works_section {
-  background-color: var(--black);
-}
-
-.teams {
-  font-size: 1.2rem;
-  color: var(--gray);
-  font-family: var(--font);
-  font-weight: 400;
-  margin-bottom: 2rem;
-}
