+++ src/Component/modals/video/index.module.css
@@ -1,37 +1,35 @@
 .modal {
   position: fixed;
   top: 0;
+  left: 0;
   bottom: 0;
   right: 0;
-  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(189, 189, 189, 0.5);
+  z-index: 555;
 }
 .modal__container {
-  display: inline-flex;
-  position: relative;
-}
-
-.modal__container source {
   position: fixed;
-  width: 100%;
-  height: 100vh;
+  width: 1228px;
+  height: 691px;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
   z-index: 999;
-  margin: auto !important;
 }
-/* .modal__video source {
-  position: fixed;
-   top: 0%;
-  left: 0;
+
+.modal__container video {
   width: 100%;
   height: 100%;
-  transform: translate(-50%, -50%); } */
+  object-fit: contain;
+}
 
 .modal__iconbutton {
   position: absolute;
-  top: 0;
-  right: 0;
+  top: -27px;
+  right: -75px;
   font-size: 3rem;
   z-index: 99999;
+  color: var(--black);
 }
