+++ src/Component/teamcards/index.module.css
@@ -49,3 +49,10 @@
 .card__button {
   padding: 0rem 13rem 0 0;
 }
+.teams {
+  font-size: 1.2rem;
+  color: var(--gray);
+  font-family: var(--font);
+  font-weight: 400;
+  margin-bottom: 2rem;
+}
