+++ src/Component/ui/featureheading/index.css
@@ -27,10 +27,12 @@
 
 .feature__bigfont {
   font-size: 2rem;
+  margin-bottom: 1rem;
 }
 
 .feature__smallfont {
   font-size: 1.4rem;
+  margin-bottom: 2rem;
 }
 .feature__biggerfont {
   font-size: 4.4rem;
