+++ src/Component/ui/heroheading/index.jsx
