+++ src/pages/home/index.css
@@ -1,0 +1,3 @@
+.works_section {
+  background-color: var(--black);
+}
